\name{CLMC-package}
\alias{CLMC-package}
\alias{CLMC}
\docType{package}
\title{
\packageTitle{CLMC}
}
\description{
\packageDESCRIPTION{CLMC}
Multiple hypothesis testing, based on composite likelihood inference for clustered data, with common covariance structure across the clusters. The critical value is based on multivariate normal quantile, using the R-package multcomp, by Hothorn, et.al.
}
\author{Mahdis Azadbakhsh,
Xin Gao,
Hanna Jankowski.
}

\references{
Azadbakhsh,M.,  Gao,X., Jankowski,H. (2016)
{Multiple Comparisons Using Composite Likelihood in Clustered Data}. Submitted.

Hothorn, T., Bretz, F., Westfall, P., Heiberger, R. M., Schutzenmeister, A.(2010)
{ multcomp: Simultaneous Inference for General Linear Hypotheses}, R package version 1.1-7, http://CRAN.R-project.org/package=multcomp
}
